Refactor api module
