+++ .gitignore
@@ -13,7 +13,7 @@ bin
 *.out
 
 # Kubernetes Generated files - skip generated files, except for vendored files
-
+vendor/
 !vendor/**/zz_generated.*
 
 # editor and IDE paraphernalia
