+++ pkg/apis/helmrelease/v2beta1/helmrelease_types.go
@@ -783,7 +783,6 @@ const (
 )
 
 // +genclient
-// +genclient:Namespaced
 // +kubebuilder:object:root=true
 // +kubebuilder:resource:shortName=hr
 // +kubebuilder:subresource:status
